refactor: rename toolset name->slug, remove dead code, deduplicate types
- Rename `name` to `slug` across backend (Rust) and frontend (TypeScript)
  for toolset instance identifiers (ToolsetRow, CreateToolsetRequest,
  UpdateToolsetRequest, ToolsetResponse, Toolset)
- Consolidate database migration 0007 to use `slug` column directly
- Remove dead code: builtin_tool_definitions(), list_all_tool_definitions(),
  is_production parameter, ToolsetTypeResponse, ToolInstanceInfo
- Deduplicate types: reuse Toolset and ToolTypeReviewInfo from ts-client
  instead of local redefinitions in useAppAccessRequests
- Replace ToolsetTypeResponse with ToolsetDefinition in frontend
- Update OpenAPI spec and regenerate TypeScript client types
- Fix bug in new/page.tsx: form.setValue was setting slug to display name

+++ crates/llama_server_proc/src/server.rs
@@ -297,7 +297,7 @@ mod tests {
       .model(PathBuf::from("/path/to/model"))
       .alias("test-alias".to_string())
       .server_args(server_args)
-      .port(12345 as u16)
+      .port(12345_u16)
       .build()
       .unwrap();
